+++ dig_core/lib/src/domain/extension/balance_extension.dart
@@ -1,0 +1,7 @@
+import 'package:dig_core/dig_core.dart';
+
+extension BalanceExtension on Balance {
+  num getToken([num ratio = 1000000]) {
+    return (double.tryParse(amount) ?? 0) / ratio;
+  }
+}
