--- dig_core/lib/src/domain/extension/double_extension.dart
@@ -1,3 +1,0 @@
-extension DoubleExtension on double {
-  double toPrecision(int n) => double.tryParse(toStringAsFixed(n)) ?? 0;
-}
